- Refresh the availability info on load. And every 5 seconds. - AND ANIMATE IT -- TODO: stop expanding =(

+++ Drugs!/CoreLogic/Models/MedicineModelDefinitions.swift
@@ -12,7 +12,7 @@ public struct Drug: EquatableFileStorable {
 
     let drugName: String
     let ingredients: [Ingredient]
-    let hourlyDoseTime: Int
+    let hourlyDoseTime: Double
     
     static func blank(_ name: String? = nil) -> Drug {
         Drug(
